requirements.txt
altgraph==0.17.4
attrs==23.2.0
Babel==2.15.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.7
customtkinter==5.2.2
darkdetect==0.8.0
exceptiongroup==1.2.1
Flask==3.0.3
fpdf==1.7.2
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
librouteros==3.2.1
MarkupSafe==2.1.5
MouseInfo==0.1.3
mysql-connector==2.2.9
mysql-connector-python==8.4.0
outcome==1.3.0.post0
packaging==24.0
paramiko==3.4.0
pillow==10.3.0
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
pyinstaller==6.7.0
pyinstaller-hooks-contrib==2024.6
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python3-xlib==0.15
pytweening==1.2.0
pywhatkit==5.4
requests==2.32.3
scapy==2.5.0
schedule==1.2.2
selenium==4.21.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
tkcalendar==1.6.1
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.0
urllib3==2.2.1
Werkzeug==3.0.3
wikipedia==1.4.0
wsproto==1.2.0
